split base class to a separate file
