+++ src/postgres_mcp/dta/__init__.py
@@ -1,16 +1,16 @@
 from pglast import parse_sql
 
 from ..artifacts import ExplainPlanArtifact
-from .dta_calc import MAX_NUM_DTA_QUERIES_LIMIT
 from .dta_calc import ColumnCollector
 from .dta_calc import ConditionColumnCollector
 from .dta_calc import DatabaseTuningAdvisor
-from .dta_calc import DTASession
 from .dta_calc import Index
 from .dta_tools import DTATool
+from .index_opt_base import MAX_NUM_INDEX_TUNING_QUERIES
+from .index_opt_base import DTASession
 
 __all__ = [
-    "MAX_NUM_DTA_QUERIES_LIMIT",
+    "MAX_NUM_INDEX_TUNING_QUERIES",
     "ColumnCollector",
     "ConditionColumnCollector",
     "DTASession",
