+++ src/postgres_mcp/dta/dta_tools.py
@@ -10,9 +10,9 @@ import humanize
 from ..artifacts import ExplainPlanArtifact
 from ..artifacts import calculate_improvement_multiple
 from ..sql import SqlDriver
-from .dta_calc import DTASession
 from .dta_calc import IndexConfig
 from .dta_calc import IndexTuningBase
+from .index_opt_base import DTASession
 
 logger = logging.getLogger(__name__)
